dfghjkl;oiuytrdfghjk

+++ include/cNES/cpu.h
@@ -61,7 +61,7 @@ void CPU_Destroy(CPU* cpu);
 void CPU_Interupt(CPU* cpu);
 void CPU_NMI(CPU* cpu);
 
-void CPU_SetFlag(CPU* cpu, uint8_t flag, int value);
+void CPU_SetFlag(CPU* cpu, uint8_t flag, uint8_t value);
 uint8_t CPU_GetFlag(CPU* cpu, uint8_t flag);
 
 #endif // CPU_H
